在systemd线程中执行fsinit成功读取超级块内容 修复唤醒丢失的问题 1. void sleep(void* chan, struct spinlock* lock) {   struct task* t = mytask();   acquire_spin(&t->lock);   if (lock)     release_spin(lock); ...   不能提前释放lock,以disk-lock为例,在fsinit时会触发sleep,   如果sleep马上释放lock,由于release_spin的开中断判断,在执行acquire_spin(&t->lock)前   cpu处于开中断状态,如果硬盘发生中断会打断sleep去执行wakeup,但此时线程状态还没有设置为sleep   导致唤醒丢失 }
2.
void task_schedule(){
  ...
  c->cur_task = NULL;     //! 不要在释放线程锁后置空,可能会被中断
  release_spin(&t->lock); //`
  ...
}

void wakeup(void* chan){
  .....
if (t != mytask()) {
      acquire_spin(&t->lock);
      if (t->state == SLEEP && t->chan == chan)
        t->state = READY;
      release_spin(&t->lock);
    }
}

~·

+++ kernel/type.h
@@ -13,7 +13,7 @@ typedef long long i64;
 typedef unsigned char u8;
 typedef unsigned short u16;
 typedef unsigned int u32;
-typedef unsigned long long u64;
+typedef unsigned long u64;
 
 typedef u64 pte_t;
 typedef pte_t* pagetable_t;
